+++ eox_tagging/api/urls.py
@@ -6,5 +6,5 @@ from django.conf.urls import include, url
 from eox_tagging.api.v1.routers import router
 
 urlpatterns = [
-    url(r'api/v1/', include(router.urls)),
+    url(r'v1/', include(router.urls)),
 ]