+++ eox_tagging/api/v1/fields.py
@@ -4,6 +4,7 @@ from rest_framework import serializers
 
 class EnumField(serializers.ChoiceField):
     """Serializer field for choices."""
+
     def __init__(self, enum, **kwargs):
         self.enum = enum
         kwargs['choices'] = [(e.value, e.name) for e in enum]
@@ -12,14 +13,13 @@ class EnumField(serializers.ChoiceField):
     def to_representation(self, obj):  # pylint: disable=arguments-differ
         """Function that helps with choice serialization."""
         try:
-            value = obj.name
-            return value
+            return getattr(obj, 'name', self.enum(obj).name)
         except AttributeError:
             return self.enum(obj).name
 
     def to_internal_value(self, data):
         """Function that helps with choice deserialization."""
         try:
-            return self.enum[data]
+            return self.enum[data.upper()]
         except KeyError:
-            self.fail('invalid_choice', input=data)
+            return None