+++ eox_tagging/urls.py
@@ -4,5 +4,5 @@ eox_tagging URL Configuration
 from django.conf.urls import include, url
 
 urlpatterns = [
-    url(r'', include('eox_tagging.api.urls')),
+    url(r'api/', include('eox_tagging.api.urls')),
 ]