+++ CHANGELOG.rst
@@ -9,8 +9,12 @@ Change Log
 
    This project adheres to Semantic Versioning (http://semver.org/).
 .. There should always be an "Unreleased" section for changes pending release.
+
 Unreleased
 ----------
+
+- REST API to create, get, filter and delete tags.
+- New filters in Tag queryset.
 
 * First PyPI release.